% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{arma_mm}
\alias{arma_mm}
\title{Matrix Product in RcppArmadillo.}
\usage{
arma_mm(m, m2)
}
\arguments{
\item{m}{numeric matrix}

\item{m2}{numeric matrix}
}
\value{
matrix product of m and m2
}
\description{
Matrix Product in RcppArmadillo.
}
