% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{solve_mm}
\alias{solve_mm}
\title{Matrix Inversion in RcppArmadillo.}
\usage{
solve_mm(Am)
}
\arguments{
\item{Am}{numeric matrix}
}
\value{
Inversion of Am
}
\description{
Matrix Inversion in RcppArmadillo.
}
